Starting on tests and moved packages
