+++ android/test/AndroidAccelerationStatusTest.java
@@ -1,0 +1,9 @@
+import cg.group4.GdxTestRunner;
+import org.junit.runner.RunWith;
+
+/**
+ * JUnit test class that tests the contents of the AndroidAccelerationStatus.java class
+ */
+@RunWith(GdxTestRunner.class)
+public class AndroidAccelerationStatusTest {
+}
