+++ core/src/cg/group4/game_logic/stroll/Stroll.java
@@ -4,8 +4,7 @@ import cg.group4.game_logic.StandUp;
 import cg.group4.game_logic.stroll.events.StrollEvent;
 import cg.group4.game_logic.stroll.events.TestStrollEvent;
 import cg.group4.game_logic.stroll.events.fishevent.FishingStrollEvent;
-import cg.group4.sensor.AccelerationState;
-import cg.group4.sensor.AccelerationStatus;
+import cg.group4.util.sensors.AccelerationState;
 import cg.group4.util.subscribe.Subject;
 import cg.group4.util.timer.Timer;
 import cg.group4.util.timer.TimerStore;
