--- core/src/cg/group4/sensor/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package containing the logic behind the sensors.
- */
-package cg.group4.sensor;
